requirements.txt
aiohttp-cors==0.7.0
airportsdata==20241001
astor==0.8.1
blake3==1.0.1
cachetools==5.5.0
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
colorful==0.5.6
compressed-tensors==0.8.1
datasets==2.19.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
einops==0.8.0
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.3.1
fuzzywuzzy==0.18.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
httptools==0.6.4
httpx==0.27.2
huggingface_hub==0.24.6
idna==3.7
importlib-metadata==8.5.0
interegular==0.3.3
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
lark==1.2.2
linkify-it-py==2.0.3
lm-format-enforcer==0.10.9
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
msgpack==1.1.0
msgspec==0.19.0
nest-asyncio==1.6.0
nvidia-ml-py==12.560.30
openai==1.52.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
outlines==0.1.11
outlines-core==0.1.26
pillow==10.4.0
platformdirs==4.3.6
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycountry==24.6.1
Pygments==2.19.1
python-dotenv==1.0.1
pyzmq==26.2.0
ray==2.40.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
starlette==0.41.3
sympy==1.13.1
textual==1.0.0
tiktoken==0.7.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.45.2
typing-extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.28.1
vllm==0.6.6.post1
wandb==0.19.2
watchfiles==1.0.4
websockets==14.1
wrapt==1.17.0
zipp==3.21.0
